+++ operator/pkg/resources/statsd/configmap.go
@@ -1,17 +1,15 @@
-// Package statsd contains k8s resources required for statsd
+// Package statsd is deprecated. StatsD support was deprecated in AIStore
+// in v3.28 and dropped in v4.0. Resource-name helpers are maintained here
+// for cleanup of past deployments.
 /*
- * Copyright (c) 2021, NVIDIA CORPORATION. All rights reserved.
+ * Copyright (c) 2021-2026, NVIDIA CORPORATION. All rights reserved.
  */
 package statsd
 
 import (
 	aisv1 "github.com/ais-operator/api/v1beta1"
-	"github.com/ais-operator/pkg/resources/ownerref"
 	"k8s.io/apimachinery/pkg/types"
-	corev1ac "k8s.io/client-go/applyconfigurations/core/v1"
 )
-
-const ConfigFile = "statsd.json"
 
 func ConfigMapName(ais *aisv1.AIStore) string {
 	return ais.Name + "-statsd"
@@ -22,15 +20,4 @@ func ConfigMapNSName(ais *aisv1.AIStore) types.NamespacedName {
 		Name:      ConfigMapName(ais),
 		Namespace: ais.Namespace,
 	}
-}
-
-func NewStatsDCM(ais *aisv1.AIStore) *corev1ac.ConfigMapApplyConfiguration {
-	return corev1ac.ConfigMap(ConfigMapName(ais), ais.Namespace).
-		WithOwnerReferences(ownerref.NewControllerRef(ais)).
-		WithData(map[string]string{
-			ConfigFile: `{
-				"graphiteHost": "",
-				"graphitePort": 2003
-			}`,
-		})
 }